+++ examples/jvm/src/main/scala/zio/examples/CausalProfilerToyExample.scala
@@ -1,12 +1,13 @@
 package zio.examples
 
 import zio._
+import zio.profiling._
 
 object CausalProfilerToyExample extends ZIOAppDefault {
 
   def run =
     CausalProfiler
-      .profile(100) {
+      .profile(ProfilerConfig.Default.copy(iterations = 100)) {
         val io = for {
           _    <- CausalProfiler.progressPoint("iteration start")
           short = ZIO.succeed(Thread.sleep(80))
