+++ profiling/jvm/src/main/scala/zio/profiling/ExperimentResult.scala
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2021 John A. De Goes and the ZIO Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package zio.profiling
+
+import zio._
+
+final case class ExperimentResult(
+  selected: ZTraceElement,
+  speedup: Float,
+  duration: Long,
+  selectedSamples: Long,
+  throughputData: List[ThroughputData]
+) {
+
+  lazy val render: List[String] =
+    s"experiment\tselected=$selected\tspeedup=$speedup\tduration=$duration\tselected-samples=$selectedSamples" :: throughputData
+      .map(_.render)
+}
