+++ profiling/jvm/src/main/scala/zio/profiling/FiberState.scala
@@ -1,0 +1,39 @@
+/*
+ * Copyright 2021 John A. De Goes and the ZIO Contributors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package zio.profiling
+
+import zio.ZTraceElement
+
+import java.util.concurrent.atomic.AtomicLong
+
+private final class FiberState(
+  val localDelay: AtomicLong,
+  @volatile var location: ZTraceElement,
+  @volatile var suspended: Boolean,
+  @volatile var preSuspendGlobalDelay: Long
+)
+
+private object FiberState {
+  
+  def initial(localDelay: Long) =
+    new FiberState(
+      new AtomicLong(localDelay),
+      null.asInstanceOf[ZTraceElement],
+      false,
+      0
+    )
+}
